third_party/blink/web_tests/fast/events/key-events-in-editable-gridbox-expected.txt

PASS targetDiv.innerText is "TEST"
PASS targetDiv.innerHTML is "<br>"
PASS targetDiv.innerText is "TEST"
PASS targetDiv.innerText is "TEST"
PASS targetDiv.innerHTML is "<br>"
PASS targetDiv.innerText is "TEST"
PASS targetDiv.innerText is "TEST"
PASS targetDiv.innerHTML is "<br>"
PASS targetDiv.innerText is "TEST"
PASS successfullyParsed is true

TEST COMPLETE

TEST
TEST
TEST
